Update base class for CheckBoxImages component
The CheckBoxImages component was previously extending the Radio class from the Filament Forms package. In this update, the base class has been changed and it now extends the Checkbox class instead. This change is expected to bring the more consistent behavior expected of a checkbox, as indicated by its name.